Adds a bit of clarity to where the stability fee goes.
This referenced `Buy&Burn` as though it was a very precisely defined thing that only applies to Maker.  In reality, Buy&Burn and Create&Sell are concepts used throughout the token space, not just with regards to Maker and not specifically related to MKR.  The way this was worded previously, it made it sound like DAI was subject to the Buy&Burn process (which it is when liquidating a CDP) when in reality it is MKR that is being bought (with DAI) and then burned.

The glossary should also be updated to define `Buy&Burn` as a more generic process, which _may_ be applied to MKR in various situations (though it also is applied to DAI within Maker).  Right now this document appears to overload the term `Buy&Burn` to be specific to MKR which is just confusing.